+++ src/components/admin/settings/service-types/ServiceTypeList.tsx
@@ -115,7 +115,7 @@ export default function ServiceTypeList({ token }: ServiceTypeListProps) {
     <div className="p-4 mx-auto">
       <div className="flex items-center gap-4 mb-4">
         <SearchBar 
-          onSearch={handleSearch} 
+          onSearch={()=>{handleSearch("")}} 
           placeholder="Buscar tipo de servicio..." 
         />
       </div>
